requirements.txt
# Google Cloud Text-to-Speech
google-cloud-texttospeech>=2.16.0

# Google Agent Development Kit (ADK)
google-adk>=0.1.0

# Google Assistant / Agent SDK
google-assistant-sdk>=0.1.0
google-assistant-context>=0.1.0
google-assistant-skill>=0.1.0

# pdf reading
streamlit>=1.28.0
PyMuPDF>=1.24.0
pymupdf4llm>=0.1.7
Pillow>=10.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
python-docx>=1.1.0

# AI & ML
google-generativeai>=0.7.0